requirements.txt
boto3==1.4.7
botocore==1.7.7
certifi==2017.7.27.1
chardet==3.0.4
docutils==0.14
future==0.16.0
idna==2.6
jmespath==0.9.3
oauthlib==2.0.3
python-dateutil==2.6.1
python-twitter==3.3
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
s3transfer==0.1.11
six==1.10.0
urllib3==1.22
